# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * InternTrain API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface StudentRegistration {
  name: string;
  email: string;
  password: string;
  studentId: string;
}

export interface StudentLogin {
  email: string;
  password: string;
}

export interface AdminLogin {
  email: string;
  password: string;
}

export interface StudentProfile {
  id: number;
  name: string;
  email: string;
  studentId: string;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  student: StudentProfile;
}

export interface AdminAuthResponse {
  token: string;
}

export interface VideoWatched {
  videoId: string;
  sectionId: string;
}

export interface QuizSubmission {
  sectionId: string;
  score: number;
  total: number;
}

export interface SectionProgress {
  sectionId: string;
  watchedVideos: string[];
  /** @nullable */
  quizScore: number | null;
  quizPassed: boolean;
  certificateEarned: boolean;
  /** @nullable */
  certificateDate: string | null;
}

export interface StudentProgress {
  sections: SectionProgress[];
}

export interface Certificate {
  id: number;
  sectionId: string;
  awardedAt: string;
}

export interface SectionStat {
  sectionId: string;
  title: string;
  studentsStarted: number;
  studentsCompleted: number;
  avgScore: number;
}

export interface AdminStats {
  totalStudents: number;
  totalVideosWatched: number;
  totalQuizzesPassed: number;
  totalCertificates: number;
  completionRate: number;
  sectionStats: SectionStat[];
}

export interface AdminStudent {
  id: number;
  name: string;
  email: string;
  studentId: string;
  createdAt: string;
  progress: StudentProgress;
}

export interface AdminCertificate {
  id: number;
  sectionId: string;
  awardedAt: string;
  student: StudentProfile;
}

export type RegisterStudent400 = {
  error: string;
};

export type RegisterStudent409 = {
  error: string;
};

export type LoginStudent401 = {
  error: string;
};

export type GetMyProgress401 = {
  error: string;
};

export type MarkVideoWatched401 = {
  error: string;
};

export type SubmitQuiz401 = {
  error: string;
};

export type GetMyCertificates401 = {
  error: string;
};

export type LoginAdmin401 = {
  error: string;
};

export type GetAdminStudent404 = {
  error: string;
};

export type DeleteAdminStudent404 = {
  error: string;
};