# lib/api-zod/src/generated/types/adminCertificate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * InternTrain API specification
 * OpenAPI spec version: 0.1.0
 */
import type { StudentProfile } from './studentProfile';

export interface AdminCertificate {
  id: number;
  sectionId: string;
  awardedAt: string;
  student: StudentProfile;
}