# lib/api-zod/src/generated/types/adminLogin.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * InternTrain API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AdminLogin {
  email: string;
  password: string;
}