# lib/api-zod/src/generated/types/adminStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * InternTrain API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SectionStat } from './sectionStat';

export interface AdminStats {
  totalStudents: number;
  totalVideosWatched: number;
  totalQuizzesPassed: number;
  totalCertificates: number;
  completionRate: number;
  sectionStats: SectionStat[];
}