# lib/api-zod/src/generated/types/certificate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * InternTrain API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Certificate {
  id: number;
  sectionId: string;
  awardedAt: string;
}