# lib/api-zod/src/generated/types/quizSubmission.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * InternTrain API specification
 * OpenAPI spec version: 0.1.0
 */

export interface QuizSubmission {
  sectionId: string;
  score: number;
  total: number;
}