# lib/api-zod/src/generated/types/sectionProgress.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * InternTrain API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SectionProgress {
  sectionId: string;
  watchedVideos: string[];
  /** @nullable */
  quizScore: number | null;
  quizPassed: boolean;
  certificateEarned: boolean;
  /** @nullable */
  certificateDate: string | null;
}