# lib/api-zod/src/generated/types/sectionStat.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * InternTrain API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SectionStat {
  sectionId: string;
  title: string;
  studentsStarted: number;
  studentsCompleted: number;
  avgScore: number;
}