# lib/api-zod/src/generated/types/studentLogin.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * InternTrain API specification
 * OpenAPI spec version: 0.1.0
 */

export interface StudentLogin {
  email: string;
  password: string;
}