# lib/api-zod/src/generated/types/studentProfile.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * InternTrain API specification
 * OpenAPI spec version: 0.1.0
 */

export interface StudentProfile {
  id: number;
  name: string;
  email: string;
  studentId: string;
  createdAt: string;
}