# lib/api-zod/src/generated/types/studentRegistration.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * InternTrain API specification
 * OpenAPI spec version: 0.1.0
 */

export interface StudentRegistration {
  name: string;
  email: string;
  password: string;
  studentId: string;
}